+++ front-end/src/app/services/icons/icon.service.ts
@@ -1,4 +1,5 @@
-import { Injectable } from '@angular/core';
+import { Injectable, PLATFORM_ID, Inject, Component, OnInit, HostListener } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
 import { TunnelService } from '../tunnel/tunnel.service';
 import { LoaderService } from '../loading-service/loader.service';
 import { BehaviorSubject } from 'rxjs';
@@ -20,6 +21,7 @@ export class IconService {
   private iconData: IconData[];
 
   constructor(
+    @Inject(PLATFORM_ID) private platformId: Object,
     loader: LoaderService,
     tunnelService: TunnelService
   ) {
@@ -29,10 +31,12 @@ export class IconService {
     })
 
     loader.startLoadingAnimation();
-    tunnelService.getPuzzlerIcons().subscribe(data => {
-      this.iconDataObservable.next(data);
-      loader.stopLoadingAnimation();
-    });
+    if (isPlatformBrowser(this.platformId)) {
+      tunnelService.getPuzzlerIcons().subscribe(data => {
+        this.iconDataObservable.next(data);
+        loader.stopLoadingAnimation();
+      });
+    }
   }
 
   getIconAccentColor(iconID: number) {
@@ -46,7 +50,7 @@ export class IconService {
   }
 
   getIconImagePath(iconID: number) {
-    if (!this.hasData) return 'puzzle-hub-profile-0.png';
+    if (this.iconData.length === 0) { return 'puzzle-hub-profile-0.png';}
     return this.getIcon(iconID).IconPath;
   }
 
@@ -56,9 +60,5 @@ export class IconService {
         return this.iconData[i];
       }
     }
-  }
-
-  get hasData() {
-    return this.iconData.length > 0;
   }
 }
