+++ front-end/src/sass/pages/_main-menu.scss
@@ -4,13 +4,6 @@
   background-color: $bg-color;
 }
 
-.jumbotron {
-  background-color: #2f343a;
-  margin-top: 2rem;
-  border-radius: 10px;
-}
-
-
 .flex-container {
   display: flex;
   flex-flow: row wrap;
