+++ front-end/src/styles.css
@@ -43,6 +43,10 @@ html {
 }
 
 * {
+  -webkit-user-select: none; /* Safari */        
+  -moz-user-select: none; /* Firefox */
+  -ms-user-select: none; /* IE10+/Edge */
+  user-select: none; /* Standard */
   font-family: 'Open Sans', sans-serif;
 }
