+++ server-files/nginx.conf
@@ -4,24 +4,10 @@
 	server 127.0.0.1:1235;
     }
 
-    upstream beta {
-	server 127.0.0.1:1236;
-    }
-
     server {
         listen       80;
         server_name  puzzlehub.io;
 	server_name  www.puzzlehub.io;
-
-        include "/opt/bitnami/nginx/conf/bitnami/phpfastcgi.conf";
-	return 301 https://$host$request_uri;
-
-        include "/opt/bitnami/nginx/conf/bitnami/bitnami-apps-prefix.conf";
-    }
-
-    server {
-        listen       80;
-	server_name  beta.puzzlehub.io;
 
         include "/opt/bitnami/nginx/conf/bitnami/phpfastcgi.conf";
 	return 301 https://$host$request_uri;
@@ -73,29 +59,6 @@
        location /api {
                include uwsgi_params;
                uwsgi_pass unix:/home/bitnami/api/uwsgi.sock;
-       }
-    }
-
-    server {
-       listen       443 ssl;
-       server_name  beta.puzzlehub.io;
-
-       ssl_certificate      /etc/letsencrypt/live/puzzlehub.io/fullchain.pem;
-       ssl_certificate_key  /etc/letsencrypt/live/puzzlehub.io/privkey.pem;
-
-       ssl_session_cache    shared:SSL:1m;
-       ssl_session_timeout  5m;
-
-       ssl_ciphers  HIGH:!aNULL:!MD5;
-       ssl_prefer_server_ciphers  on;
-	
-       location / {
-                proxy_pass http://beta;
-                proxy_http_version 1.1;
-                proxy_set_header Upgrade $http_upgrade;
-                proxy_set_header Connection 'upgrade';
-                proxy_set_header Host $host;
-                proxy_cache_bypass Host $host;
        }
     }
